Fix unit tests
* Separated the test
  files for the
  appropriate platforms.
