+++ hadoop-hdfs-project/hadoop-hdfs-native-client/src/main/native/libhdfspp/tests/x-platform/utils_nix_test.cc
@@ -16,54 +16,20 @@
  * limitations under the License.
  */
 
-#include "x-platform/utils.h"
-
 #include <gtest/gtest.h>
 
 #include <string>
 
-TEST(XPlatformUtils, BasenameEmpty) {
-  const std::string expected(".");
-  const auto actual = XPlatform::Utils::Basename("");
-  EXPECT_EQ(expected, actual);
-}
-
-TEST(XPlatformUtils, BasenameRelativePath) {
-  const std::string expected("x");
-  const auto actual = XPlatform::Utils::Basename("x");
-  EXPECT_EQ(expected, actual);
-}
+#include "x-platform/utils.h"
 
 TEST(XPlatformUtils, BasenameRoot) {
-  const std::string win_expected(R"(\)");
-  const auto win_actual_1 = XPlatform::Utils::Basename(R"(\)");
-  EXPECT_EQ(win_expected, win_actual_1);
-
-  const auto win_actual_2 = XPlatform::Utils::Basename(R"(C:\)");
-  EXPECT_EQ(win_expected, win_actual_2);
-
   const std::string nix_expected("/");
   const auto nix_actual = XPlatform::Utils::Basename("/");
   EXPECT_EQ(nix_expected, nix_actual);
 }
 
-TEST(XPlatformUtils, BasenameSpecialFiles) {
-  const std::string current_dir_expected(".");
-  const auto current_dir_actual = XPlatform::Utils::Basename(".");
-  EXPECT_EQ(current_dir_expected, current_dir_actual);
-
-  const std::string parent_dir_expected("..");
-  const auto parent_dir_actual = XPlatform::Utils::Basename("..");
-  EXPECT_EQ(parent_dir_expected, parent_dir_actual);
-}
-
 TEST(XPlatformUtils, BasenameTrailingSlash) {
   const std::string expected("def");
-
-  const std::string win_path(R"(C:\abc\def\)");
-  const auto win_actual = XPlatform::Utils::Basename(win_path);
-  EXPECT_EQ(expected, win_actual);
-
   const std::string nix_path("/abc/def/");
   const auto nix_actual = XPlatform::Utils::Basename(nix_path);
   EXPECT_EQ(expected, nix_actual);
@@ -71,17 +37,7 @@ TEST(XPlatformUtils, BasenameTrailingSlash) {
 
 TEST(XPlatformUtils, BasenameBasic) {
   const std::string expected("def");
-
-  const std::string win_path(R"(C:\abc\def)");
-  const auto win_actual = XPlatform::Utils::Basename(win_path);
-  EXPECT_EQ(expected, win_actual);
-
   const std::string nix_path("/abc/def");
   const auto nix_actual = XPlatform::Utils::Basename(nix_path);
   EXPECT_EQ(expected, nix_actual);
-}
-
-int main(int argc, char* argv[]) {
-  ::testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
 }
